add class for caching Fixes #522
This commit adds a new class that handles Task caching. This change
pulls the logic for caching out of the TasksRepository to more closely
adhere to the Single Responsibility Principle (SRP). Now the repository
need not worry about null cache objects, iterating through the cache, or
keeping track of whether the cache is dirty or not.