Using static methods for stateless Modules
Since ApplicationModule is stateless, we could make all `@Provides` methods static. 

This will improve performance: an instance of this module is no longer required when providing these objects (you can see it in the generated code). The ApplicationComponent will call the static methods of this class directly without having to pass an instance of ApplicationModule in.